Add files via upload
Taken out the new3DLandmarks where the most probable problem lays, added initial bootstrap into the main code so we can try it with every data set. in new3DLandmarks, there is the long description before comented code with my idea to fix it.  I moved the filter from the end of processFrame into the new3DLandmarks but only checking if the point is in front of the camera - you can add another ones.